Get ReadmeEditor working
