+++ package.json
@@ -7,7 +7,7 @@
   "repository": "CONP-PCNO/conp-dats-editor",
   "main": "dist/index.js",
   "module": "dist/index.modern.js",
-  "source": "src/index.js",
+  "source": "src/index.jsx",
   "engines": {
     "node": ">=10"
   },
@@ -39,7 +39,7 @@
     "eslint-plugin-node": "^11.0.0",
     "eslint-plugin-prettier": "^3.1.1",
     "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.20.6",
+    "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-standard": "^4.0.1",
     "gh-pages": "^2.2.0",
     "jest-environment-jsdom-sixteen": "^1.0.3",
